handle circular reference with UnknownContentModel
this also fixes some lint errors
