+++ src/models/tools/tool-tile.ts
@@ -1,18 +1,17 @@
 import { cloneDeep } from "lodash";
 import { getSnapshot, Instance, SnapshotIn, SnapshotOut, types } from "mobx-state-tree";
-import { kPlaceholderToolID } from "./placeholder/placeholder-content";
 import { ITileExportOptions } from "./tool-content-info";
 import { findMetadata, ToolContentUnion } from "./tool-types";
 import { DisplayUserTypeEnum } from "../stores/user-types";
 import { uniqueId } from "../../utilities/js-utils";
 
 // import all tools so they are registered
+import "./unknown-content";
+import { kPlaceholderToolID } from "./placeholder/placeholder-content";
 import "./geometry/geometry-content";
 import "./image/image-content";
-import "./placeholder/placeholder-content";
 import "./table/table-content";
 import "./text/text-content";
-import "./unknown-content";
 import "./drawing/drawing-content";
 
 // generally negotiated with app, e.g. single column width for table
